refactor tests for new features
