+++ src/test/java/webdriver/yandex/disk/CreateNewFolderWithContextClickForLinux.java
@@ -1,6 +1,5 @@
 package webdriver.yandex.disk;
 
-import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterTest;
@@ -11,13 +10,8 @@ import webdriver.browser.TypeBrowser;
 import webdriver.screen.yandex.disk.LoginPage;
 import webdriver.screen.yandex.disk.YandexDiskPage;
 
-public class CreateNewFolderWithContextClickForLinux {
-    String sauceUserName = "d.druchkov";
-    String sauceAccessKey = "fdf2abf5-fc5e-4a81-899b-5a9e7f230ed5";
-    String url = "https://" + sauceUserName + ":" + sauceAccessKey + "@ondemand.eu-central-1.saucelabs.com:443/wd/hub";
-    WebDriver driver;
+public class CreateNewFolderWithContextClickForLinux extends SourceTest{
     String nameFolder = "test";
-
 
     @BeforeTest(description = "set up driver for work on linux")
     public void setup() {
@@ -38,10 +32,7 @@ public class CreateNewFolderWithContextClickForLinux {
     public void createNewFolderWithContextClick() {
         new LoginPage(driver)
                 .open()
-                .typeLogin("qaAutomationGomel")
-                .clickOnLogin()
-                .typePassword("gomelQa")
-                .clickOnLogin();
+                .login(user);
         YandexDiskPage pageObject = new YandexDiskPage(driver);
                 pageObject.createFolderFromContextMenu(nameFolder);
         Assert.assertTrue(pageObject.isFileExist(nameFolder) );
